+++ CHANGELOG.md
@@ -3,6 +3,7 @@ Not Yet Released
 
 ## Popup
 - Possibility to make the kanji search results more or less broaden through the target icon in the search results navbar. If target is not activated, then the result of the search will be exactly what was typed in (i.e.: searching gold will only show results with gold (results as goldfish, for example, will not be shown))
+- Added the list of kanji that were highlighted in that page to the extension popup
 
 # [Changelog v0.4.4](https://github.com/digas99/wanikani-kanji-highlighter/releases/tag/v0.4.4)
 Released on 30/07/2021